Add files via upload
Modified MainActivity.java -- because in this file there is modification of code, before transaction.commit() one line of code is added transaction.addBackToStack(String s);. This line will help to track the transactions of the tasks we performed accordingly. When user presses the back button, the previous activities are destroyed step by step. In this java file the purpose of  transaction.addBackToStack(String s); is shown.